requirements2.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.40.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==24.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.26.1
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==65.5.0
six==1.17.0
smmap==5.0.2
streamlit==1.44.0
sympy==1.14.0
tenacity==9.1.2
tesseract==0.1.3
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0


# pillow==10.3.0 
# pytesseract==0.3.10 
# protobuf==5.26.1 
# packaging==24.0